Update FoundationsDataScience_Winter2025_Group17.ipynb
It turns out that the linear regression module from sklearn wasn't added to the list of imported libraries at the top of the notebook - this meant that if all the cells were cleared, then my portion of the code wouldn't run - Whoops!
I've added that 1 line and reran my portion of the code to generate the tables.